Break out user and all shouts components
